refactor(w1): point PluginSettings importers at domain/settings (#99)
11 files now import PluginSettings/DEFAULT_SETTINGS from the new
canonical location. IBridge.ts re-export remains for one more chunk
while IBridge itself is migrated, then both are deleted together.
