+++ src/ui/composables/useSettings.ts
@@ -4,7 +4,7 @@ import { useSettingsStore } from '../stores/settingsStore'
 import { setLocale } from '../i18n'
 import type { SupportedLocale } from '../i18n'
 import { tryAsync } from '@/domain/shared/tryAsync'
-import type { PluginSettings } from '@/infrastructure/bridge/IBridge'
+import type { PluginSettings } from '@/domain/settings/PluginSettings'
 
 export function useSettings() {
   const bridge = useBridge()
